Remove out-of-bounds lookups to fix Chrome performance
Previously the code wasn’t doing bounds checking and consequently was checking the array for negative indexes. This appears to throw Chrome into some sort of unoptimised path.

We want wrap behaviour at the edges anyway too keep the system enclosed into a single ‘universe’, so this commit adds code to check to see wrap edge lookups.

The original problem is detailed in my question on StackOverflow:
http://stackoverflow.com/q/35831151/453783

The associated Chromium bug is:
https://bugs.chromium.org/p/v8/issues/detail?id=4816